requirements.txt
aiogram
aiosqlite
python-dotenv
qrcode
pillow
requests